drivers: clock_control: split esp32 driver by clock ip
Move the per-SoC clock logic out of the single clock_control_esp32.c
into two family files selected at build time: one for the RTC_CNTL
based parts (esp32, s2, s3, c2, c3) and one for the PMU/LP_CLKRST
based parts (c5, c6, h2, p4). Shared device API ops, init and the
common includes move to a shell file and a private header.

This removes the cross-series preprocessor chains that dominated the
CPU and slow-clock routines and matches how other multi-SoC clock
drivers are organized in the tree.
