+++ drivers/clock_control/CMakeLists.txt
@@ -64,6 +64,19 @@ if(CONFIG_CLOCK_CONTROL_RCAR_CPG_MSSR)
   zephyr_library_sources_ifdef(CONFIG_DT_HAS_RENESAS_R8A779G0_CPG_MSSR_ENABLED clock_control_r8a779g0_cpg_mssr.c)
 endif()
 
+if(CONFIG_CLOCK_CONTROL_ESP32)
+  zephyr_library_sources(clock_control_esp32.c)
+  zephyr_library_sources_ifdef(CONFIG_SOC_SERIES_ESP32 clock_control_esp32_rtc_cntl.c)
+  zephyr_library_sources_ifdef(CONFIG_SOC_SERIES_ESP32C2 clock_control_esp32_rtc_cntl.c)
+  zephyr_library_sources_ifdef(CONFIG_SOC_SERIES_ESP32C3 clock_control_esp32_rtc_cntl.c)
+  zephyr_library_sources_ifdef(CONFIG_SOC_SERIES_ESP32C5 clock_control_esp32_pmu.c)
+  zephyr_library_sources_ifdef(CONFIG_SOC_SERIES_ESP32C6 clock_control_esp32_pmu.c)
+  zephyr_library_sources_ifdef(CONFIG_SOC_SERIES_ESP32H2 clock_control_esp32_pmu.c)
+  zephyr_library_sources_ifdef(CONFIG_SOC_SERIES_ESP32P4 clock_control_esp32_pmu.c)
+  zephyr_library_sources_ifdef(CONFIG_SOC_SERIES_ESP32S2 clock_control_esp32_rtc_cntl.c)
+  zephyr_library_sources_ifdef(CONFIG_SOC_SERIES_ESP32S3 clock_control_esp32_rtc_cntl.c)
+endif()
+
 # zephyr-keep-sorted-start
 zephyr_library_sources_ifdef(CONFIG_CLOCK_CONTROL_ADSP clock_control_adsp.c)
 zephyr_library_sources_ifdef(CONFIG_CLOCK_CONTROL_AGILEX5 clock_control_agilex5.c)
@@ -79,7 +92,6 @@ zephyr_library_sources_ifdef(CONFIG_CLOCK_CONTROL_BOUFFALOLAB_BL60X clock_contro
 zephyr_library_sources_ifdef(CONFIG_CLOCK_CONTROL_BOUFFALOLAB_BL61X clock_control_bl61x.c)
 zephyr_library_sources_ifdef(CONFIG_CLOCK_CONTROL_BOUFFALOLAB_BL70X_L clock_control_bl70x_l.c)
 zephyr_library_sources_ifdef(CONFIG_CLOCK_CONTROL_BOUFFALOLAB_BL808 clock_control_bl808.c)
-zephyr_library_sources_ifdef(CONFIG_CLOCK_CONTROL_ESP32 clock_control_esp32.c)
 zephyr_library_sources_ifdef(CONFIG_CLOCK_CONTROL_FIXED_RATE_CLOCK clock_control_fixed_rate.c)
 zephyr_library_sources_ifdef(CONFIG_CLOCK_CONTROL_FOCALTECH_FT9001 clock_control_ft9001.c)
 zephyr_library_sources_ifdef(CONFIG_CLOCK_CONTROL_GD32 clock_control_gd32.c)
