+++ drivers/clock_control/clock_control_esp32.c
@@ -7,102 +7,8 @@
 
 #define DT_DRV_COMPAT espressif_esp32_clock
 
-#define CPU_RESET_REASON RTC_SW_CPU_RESET
+#include "clock_control_esp32_priv.h"
 
-#if defined(CONFIG_SOC_SERIES_ESP32)
-#define DT_CPU_COMPAT espressif_xtensa_lx6
-#undef CPU_RESET_REASON
-#define CPU_RESET_REASON SW_CPU_RESET
-#include <zephyr/dt-bindings/clock/esp32_clock.h>
-#include <esp32/rom/rtc.h>
-#include <soc/dport_reg.h>
-#include <soc/i2s_reg.h>
-#elif defined(CONFIG_SOC_SERIES_ESP32S2)
-#define DT_CPU_COMPAT espressif_xtensa_lx7
-#include <zephyr/dt-bindings/clock/esp32s2_clock.h>
-#include <esp32s2/rom/rtc.h>
-#include <soc/dport_reg.h>
-#include <soc/i2s_reg.h>
-#elif defined(CONFIG_SOC_SERIES_ESP32S3)
-#define DT_CPU_COMPAT espressif_xtensa_lx7
-#include <zephyr/dt-bindings/clock/esp32s3_clock.h>
-#include <esp32s3/rom/rtc.h>
-#include <soc/dport_reg.h>
-#elif defined(CONFIG_SOC_SERIES_ESP32C2)
-#define DT_CPU_COMPAT espressif_riscv
-#include <zephyr/dt-bindings/clock/esp32c2_clock.h>
-#include <esp32c2/rom/rtc.h>
-#elif defined(CONFIG_SOC_SERIES_ESP32C3)
-#define DT_CPU_COMPAT espressif_riscv
-#include <zephyr/dt-bindings/clock/esp32c3_clock.h>
-#include <esp32c3/rom/rtc.h>
-#elif defined(CONFIG_SOC_SERIES_ESP32C5)
-#define DT_CPU_COMPAT espressif_riscv
-#include <zephyr/dt-bindings/clock/esp32c5_clock.h>
-#include <soc/lp_clkrst_reg.h>
-#include <soc/pmu_reg.h>
-#include <soc/regi2c_dig_reg.h>
-#include <regi2c_ctrl.h>
-#include <esp32c5/rom/rtc.h>
-#include <soc/dport_access.h>
-#include <hal/clk_tree_ll.h>
-#include <esp_private/esp_pmu.h>
-#include <modem/modem_syscon_struct.h>
-#elif defined(CONFIG_SOC_SERIES_ESP32C6)
-#define DT_CPU_COMPAT espressif_riscv
-#include <zephyr/dt-bindings/clock/esp32c6_clock.h>
-#include <soc/lp_clkrst_reg.h>
-#include <soc/pmu_reg.h>
-#include <soc/regi2c_dig_reg.h>
-#include <regi2c_ctrl.h>
-#include <esp32c6/rom/rtc.h>
-#include <soc/dport_access.h>
-#include <hal/clk_tree_ll.h>
-#include <esp_private/esp_pmu.h>
-#include <modem/modem_lpcon_struct.h>
-#elif defined(CONFIG_SOC_SERIES_ESP32H2)
-#define DT_CPU_COMPAT espressif_riscv
-#include <zephyr/dt-bindings/clock/esp32h2_clock.h>
-#include <soc/lp_clkrst_reg.h>
-#include <regi2c_ctrl.h>
-#include <esp32h2/rom/rtc.h>
-#include <soc/dport_access.h>
-#include <hal/clk_tree_ll.h>
-#include <esp_private/esp_pmu.h>
-#elif defined(CONFIG_SOC_SERIES_ESP32P4)
-#define DT_CPU_COMPAT espressif_riscv
-#include <zephyr/dt-bindings/clock/esp32p4_clock.h>
-#include <soc/lp_clkrst_reg.h>
-#include <soc/pmu_reg.h>
-#include <soc/regi2c_dig_reg.h>
-#include <regi2c_ctrl.h>
-#include <esp32p4/rom/rtc.h>
-#include <soc/dport_access.h>
-#include <hal/clk_tree_ll.h>
-#include <esp_private/esp_pmu.h>
-#endif
-
-#include <zephyr/drivers/clock_control.h>
-#include <zephyr/drivers/clock_control/esp32_clock_control.h>
-
-#include <esp_rom_caps.h>
-#include <esp_rom_serial_output.h>
-#include <esp_rom_sys.h>
-#include <hal/timg_ll.h>
-#include <hal/uart_ll.h>
-#include <soc/periph_defs.h>
-#include <soc/rtc.h>
-#include <hal/clk_gate_ll.h>
-#include <esp_private/periph_ctrl.h>
-#include <esp_private/esp_clk.h>
-#include <esp_cpu.h>
-#include <hal/regi2c_ctrl_ll.h>
-#include <hal/clk_tree_hal.h>
-#include <esp_private/esp_clk_tree_common.h>
-#include <esp_private/rtc_clk.h>
-#include "esp_clk_internal.h"
-
-#include <zephyr/irq.h>
 #include <zephyr/logging/log.h>
 LOG_MODULE_REGISTER(clock_control, CONFIG_CLOCK_CONTROL_LOG_LEVEL);
 
@@ -196,399 +102,6 @@ static int clock_control_esp32_get_rate(const struct device *dev, clock_control_
 		*rate = clk_hal_cpu_get_freq_hz();
 	}
 
-	return 0;
-}
-
-static int esp32_select_rtc_slow_clk(uint8_t slow_clk)
-{
-#if !defined(CONFIG_SOC_SERIES_ESP32C6) && !defined(CONFIG_SOC_SERIES_ESP32C5) &&                  \
-	!defined(CONFIG_SOC_SERIES_ESP32H2) && !defined(CONFIG_SOC_SERIES_ESP32P4)
-	soc_rtc_slow_clk_src_t rtc_slow_clk_src = slow_clk & RTC_CNTL_ANA_CLK_RTC_SEL_V;
-#else
-	soc_rtc_slow_clk_src_t rtc_slow_clk_src = slow_clk;
-#endif
-	uint32_t cal_val = 0;
-	/* number of times to repeat 32k XTAL calibration
-	 * before giving up and switching to the internal RC
-	 */
-	int retry_32k_xtal = 3;
-
-	/*
-	 * RTC slow clock calibration uses TIMG0 to count XTAL cycles.
-	 * Ensure TIMG0 bus clock is enabled before calibration.
-	 */
-	_timg_ll_enable_bus_clock(0, true);
-	_timg_ll_reset_register(0);
-	do {
-#if defined(CONFIG_SOC_SERIES_ESP32C2)
-		if (rtc_slow_clk_src == ESP32_RTC_SLOW_CLK_SRC_OSC_SLOW) {
-			/* external clock needs to be connected to PIN0 before it can
-			 * be used. Here we use rtc_clk_cal function to count
-			 * the number of ext clk cycles in the given number of ext clk
-			 * cycles. If the ext clk has not started up, calibration
-			 * will time out, returning 0.
-			 */
-			LOG_DBG("waiting for external clock by pin0 to start up");
-			rtc_clk_32k_enable_external();
-#else
-		if (rtc_slow_clk_src == ESP32_RTC_SLOW_CLK_SRC_XTAL32K) {
-			/* 32k XTAL oscillator needs to be enabled and running before it can
-			 * be used. Hardware doesn't have a direct way of checking if the
-			 * oscillator is running. Here we use rtc_clk_cal function to count
-			 * the number of main XTAL cycles in the given number of 32k XTAL
-			 * oscillator cycles. If the 32k XTAL has not started up, calibration
-			 * will time out, returning 0.
-			 */
-			LOG_DBG("waiting for 32k oscillator to start up");
-			if (slow_clk == ESP32_RTC_SLOW_CLK_SRC_XTAL32K) {
-				rtc_clk_32k_enable(true);
-			}
-#if !defined(CONFIG_SOC_SERIES_ESP32P4)
-			else if (slow_clk == ESP32_RTC_SLOW_CLK_32K_EXT_OSC) {
-				rtc_clk_32k_enable_external();
-			} else {
-			}
-#endif
-#endif
-			/* When CLOCK_CONTROL_ESP32_RTC_CLK_CAL_CYCLES is set to 0, clock
-			 * calibration will not be performed at startup.
-			 */
-			if (CONFIG_CLOCK_CONTROL_ESP32_RTC_CLK_CAL_CYCLES > 0) {
-#if defined(CONFIG_SOC_SERIES_ESP32C2)
-				cal_val =
-					rtc_clk_cal(CLK_CAL_32K_OSC_SLOW,
-						    CONFIG_CLOCK_CONTROL_ESP32_RTC_CLK_CAL_CYCLES);
-#else
-				cal_val =
-					rtc_clk_cal(CLK_CAL_32K_XTAL,
-						    CONFIG_CLOCK_CONTROL_ESP32_RTC_CLK_CAL_CYCLES);
-#endif
-				if (cal_val == 0) {
-					if (retry_32k_xtal-- > 0) {
-						continue;
-					}
-					LOG_ERR("32 kHz XTAL not found");
-					return -ENODEV;
-				}
-			}
-#if defined(CONFIG_SOC_SERIES_ESP32C6) || defined(CONFIG_SOC_SERIES_ESP32H2) ||                    \
-	defined(CONFIG_SOC_SERIES_ESP32P4)
-		} else if (rtc_slow_clk_src == SOC_RTC_SLOW_CLK_SRC_RC32K) {
-			rtc_clk_rc32k_enable(true);
-		}
-#elif defined(CONFIG_SOC_SERIES_ESP32C5)
-		}
-#else
-		} else if (rtc_slow_clk_src == SOC_RTC_SLOW_CLK_SRC_RC_FAST_D256) {
-			rtc_clk_8m_enable(true, true);
-		}
-#endif
-		rtc_clk_slow_src_set(rtc_slow_clk_src);
-
-#if defined(CONFIG_SOC_SERIES_ESP32C5) || defined(CONFIG_SOC_SERIES_ESP32C6) ||                    \
-	defined(CONFIG_SOC_SERIES_ESP32H2) || defined(CONFIG_SOC_SERIES_ESP32P4)
-		/*
-		 * The source enums differ per SoC (ESP32-C5 has no RC32K,
-		 * ESP32-P4 has no OSC_SLOW), so each term is computed under the
-		 * SoC guard where its enum exists.
-		 */
-		bool xpd_xtal32k = (rtc_slow_clk_src == ESP32_RTC_SLOW_CLK_SRC_XTAL32K);
-		bool xpd_rc32k = false;
-
-#if !defined(CONFIG_SOC_SERIES_ESP32P4)
-		xpd_xtal32k |= (rtc_slow_clk_src == SOC_RTC_SLOW_CLK_SRC_OSC_SLOW);
-#endif
-#if !defined(CONFIG_SOC_SERIES_ESP32C5)
-		xpd_rc32k = (rtc_slow_clk_src == SOC_RTC_SLOW_CLK_SRC_RC32K);
-#endif
-
-		pmu_lp_power_t lp_clk_power = {
-			.xpd_xtal32k = xpd_xtal32k,
-			.xpd_rc32k = xpd_rc32k,
-			.xpd_fosc = 1,
-			.pd_osc = 0,
-		};
-		pmu_ll_lp_set_clk_power(&PMU, PMU_MODE_LP_ACTIVE, lp_clk_power.val);
-#endif
-
-		if (CONFIG_CLOCK_CONTROL_ESP32_RTC_CLK_CAL_CYCLES > 0) {
-			cal_val = rtc_clk_cal(CLK_CAL_RTC_SLOW,
-					      CONFIG_CLOCK_CONTROL_ESP32_RTC_CLK_CAL_CYCLES);
-		} else {
-			const uint64_t cal_dividend = (1ULL << RTC_CLK_CAL_FRACT) * 1000000ULL;
-
-			cal_val = (uint32_t)(cal_dividend / rtc_clk_slow_freq_get_hz());
-		}
-	} while (cal_val == 0);
-
-	LOG_DBG("RTC_SLOW_CLK calibration value: %d", cal_val);
-
-	esp_clk_slowclk_cal_set(cal_val);
-
-	return 0;
-}
-
-static bool clock_init_done;
-
-static void esp32_cpu_clock_init(const struct esp32_cpu_clock_config *cpu_cfg)
-{
-	rtc_clk_config_t rtc_clk_cfg = RTC_CLK_CONFIG_DEFAULT();
-
-	rtc_clk_cfg.xtal_freq = cpu_cfg->xtal_freq;
-	rtc_clk_cfg.cpu_freq_mhz = cpu_cfg->cpu_freq;
-
-#if defined(CONFIG_SOC_SERIES_ESP32C5) || defined(CONFIG_SOC_SERIES_ESP32C6) ||                    \
-	defined(CONFIG_SOC_SERIES_ESP32H2) || defined(CONFIG_SOC_SERIES_ESP32P4)
-	_regi2c_ctrl_ll_master_enable_clock(true);
-#endif
-
-#if defined(CONFIG_SOC_SERIES_ESP32C5) || defined(CONFIG_SOC_SERIES_ESP32C6)
-	REGI2C_WRITE_MASK(I2C_DIG_REG, I2C_DIG_REG_SCK_DCAP, rtc_clk_cfg.slow_clk_dcap);
-	REGI2C_WRITE_MASK(I2C_DIG_REG, I2C_DIG_REG_ENIF_RTC_DREG, 1);
-	REGI2C_WRITE_MASK(I2C_DIG_REG, I2C_DIG_REG_ENIF_DIG_DREG, 1);
-#elif defined(CONFIG_SOC_SERIES_ESP32P4)
-	REGI2C_WRITE_MASK(I2C_DIG_REG, I2C_DIG_REG_SCK_DCAP, rtc_clk_cfg.slow_clk_dcap);
-	REGI2C_WRITE_MASK(I2C_DIG_REG, I2C_DIG_REG_FORCE_RTC_DREG, 1);
-	REGI2C_WRITE_MASK(I2C_DIG_REG, I2C_DIG_REG_FORCE_DIG_DREG, 1);
-	REGI2C_WRITE_MASK(I2C_DIG_REG, I2C_DIG_REG_XPD_RTC_REG, 0);
-	REGI2C_WRITE_MASK(I2C_DIG_REG, I2C_DIG_REG_XPD_DIG_REG, 0);
-#elif defined(CONFIG_SOC_SERIES_ESP32H2)
-	REGI2C_WRITE_MASK(I2C_PMU, I2C_PMU_OC_SCK_DCAP, rtc_clk_cfg.slow_clk_dcap);
-	REGI2C_WRITE_MASK(I2C_PMU, I2C_PMU_EN_I2C_RTC_DREG, 0);
-	REGI2C_WRITE_MASK(I2C_PMU, I2C_PMU_EN_I2C_DIG_DREG, 0);
-#else
-	REG_SET_FIELD(RTC_CNTL_REG, RTC_CNTL_SCK_DCAP, rtc_clk_cfg.slow_clk_dcap);
-	REG_SET_FIELD(RTC_CNTL_CLK_CONF_REG, RTC_CNTL_CK8M_DFREQ, rtc_clk_cfg.clk_8m_dfreq);
-#endif
-
-#if defined(CONFIG_SOC_SERIES_ESP32C5)
-	REG_SET_FIELD(LP_CLKRST_FOSC_CNTL_REG, LP_CLKRST_FOSC_DFREQ, rtc_clk_cfg.clk_8m_dfreq);
-#elif defined(CONFIG_SOC_SERIES_ESP32C6) || defined(CONFIG_SOC_SERIES_ESP32H2) ||                  \
-	defined(CONFIG_SOC_SERIES_ESP32P4)
-	REG_SET_FIELD(LP_CLKRST_FOSC_CNTL_REG, LP_CLKRST_FOSC_DFREQ, rtc_clk_cfg.clk_8m_dfreq);
-	REG_SET_FIELD(LP_CLKRST_RC32K_CNTL_REG, LP_CLKRST_RC32K_DFREQ, rtc_clk_cfg.rc32k_dfreq);
-#endif
-
-#if defined(CONFIG_SOC_SERIES_ESP32C5) || defined(CONFIG_SOC_SERIES_ESP32C6) ||                    \
-	defined(CONFIG_SOC_SERIES_ESP32H2) || defined(CONFIG_SOC_SERIES_ESP32P4)
-	/*
-	 * When PVT is enabled, esp_rtc_init() -> pmu_init() already
-	 * hands DBIAS control to PVT (DBIAS_SEL=0) and enables dynamic
-	 * tracking. Taking it back to PMU here stops PVT from updating
-	 * the RO HP/LP DBIAS_VOL fields, causing pvt_func_enable(false)
-	 * at sleep entry to read stale values (0) and write an invalid
-	 * regulator bias which brown-outs the core.
-	 */
-#if !defined(CONFIG_ESP_ENABLE_PVT)
-	uint32_t hp_cali_dbias = get_act_hp_dbias();
-	uint32_t lp_cali_dbias = get_act_lp_dbias();
-
-#if defined(CONFIG_SOC_SERIES_ESP32C5) || defined(CONFIG_SOC_SERIES_ESP32C6) ||                    \
-	defined(CONFIG_SOC_SERIES_ESP32P4)
-	SET_PERI_REG_MASK(PMU_HP_ACTIVE_HP_REGULATOR0_REG, PMU_DIG_REGULATOR0_DBIAS_SEL);
-#endif
-
-	SET_PERI_REG_BITS(PMU_HP_ACTIVE_HP_REGULATOR0_REG, PMU_HP_ACTIVE_HP_REGULATOR_DBIAS,
-			  hp_cali_dbias, PMU_HP_ACTIVE_HP_REGULATOR_DBIAS_S);
-	SET_PERI_REG_BITS(PMU_HP_MODEM_HP_REGULATOR0_REG, PMU_HP_MODEM_HP_REGULATOR_DBIAS,
-			  hp_cali_dbias, PMU_HP_MODEM_HP_REGULATOR_DBIAS_S);
-	SET_PERI_REG_BITS(PMU_HP_SLEEP_LP_REGULATOR0_REG, PMU_HP_SLEEP_LP_REGULATOR_DBIAS,
-			  lp_cali_dbias, PMU_HP_SLEEP_LP_REGULATOR_DBIAS_S);
-#endif /* !CONFIG_ESP_ENABLE_PVT */
-#endif
-
-#if defined(CONFIG_SOC_SERIES_ESP32)
-	REG_SET_FIELD(RTC_CNTL_CLK_CONF_REG, RTC_CNTL_CK8M_DIV_SEL, rtc_clk_cfg.clk_8m_div - 1);
-#elif defined(CONFIG_SOC_SERIES_ESP32C5)
-	clk_ll_rc_fast_tick_conf();
-	esp_rom_output_tx_wait_idle(0);
-#elif defined(CONFIG_SOC_SERIES_ESP32C6) || defined(CONFIG_SOC_SERIES_ESP32H2)
-	clk_ll_rc_fast_tick_conf();
-
-	esp_rom_output_tx_wait_idle(0);
-	rtc_clk_xtal_freq_update(rtc_clk_cfg.xtal_freq);
-#elif defined(CONFIG_SOC_SERIES_ESP32P4)
-	esp_rom_output_tx_wait_idle(0);
-	rtc_clk_xtal_freq_update(rtc_clk_cfg.xtal_freq);
-#else
-	rtc_clk_divider_set(rtc_clk_cfg.clk_rtc_clk_div);
-	rtc_clk_8m_divider_set(rtc_clk_cfg.clk_8m_clk_div);
-#endif
-
-#if !defined(CONFIG_SOC_SERIES_ESP32C5) && !defined(CONFIG_SOC_SERIES_ESP32C6) &&                  \
-	!defined(CONFIG_SOC_SERIES_ESP32H2) && !defined(CONFIG_SOC_SERIES_ESP32P4)
-	regi2c_ctrl_ll_i2c_reset();
-	regi2c_ctrl_ll_i2c_bbpll_enable();
-#endif
-
-#if defined(CONFIG_SOC_SERIES_ESP32S2) || defined(CONFIG_SOC_SERIES_ESP32)
-	regi2c_ctrl_ll_i2c_apll_enable();
-#endif
-
-#if !defined(CONFIG_SOC_SERIES_ESP32S2) && !defined(CONFIG_SOC_SERIES_ESP32C5) &&                  \
-	!defined(CONFIG_SOC_SERIES_ESP32P4)
-	rtc_clk_xtal_freq_update(rtc_clk_cfg.xtal_freq);
-#endif
-#if defined(CONFIG_SOC_SERIES_ESP32C6)
-	clk_ll_mspi_fast_set_hs_divider(6);
-#elif !defined(CONFIG_SOC_SERIES_ESP32C5) && !defined(CONFIG_SOC_SERIES_ESP32H2) &&                \
-	!defined(CONFIG_SOC_SERIES_ESP32P4)
-	rtc_clk_apb_freq_update(rtc_clk_cfg.xtal_freq * MHZ(1));
-#endif
-}
-
-static int esp32_cpu_clock_configure(const struct esp32_cpu_clock_config *cpu_cfg)
-{
-	rtc_cpu_freq_config_t old_config;
-	rtc_cpu_freq_config_t new_config;
-	bool ret;
-
-	if (!clock_init_done) {
-		esp32_cpu_clock_init(cpu_cfg);
-		clock_init_done = true;
-	}
-
-	esp_rom_output_tx_wait_idle(CONFIG_ESP_CONSOLE_UART_NUM);
-
-#if defined(CONFIG_SOC_SERIES_ESP32C2) || defined(CONFIG_SOC_SERIES_ESP32C5) ||                    \
-	defined(CONFIG_SOC_SERIES_ESP32C6) || defined(CONFIG_SOC_SERIES_ESP32H2) ||                \
-	defined(CONFIG_SOC_SERIES_ESP32P4)
-#if defined(CONFIG_SOC_SERIES_ESP32C5) || defined(CONFIG_SOC_SERIES_ESP32P4)
-	if (cpu_cfg->clk_src == ESP32_CPU_CLK_SRC_XTAL) {
-#else
-	if (cpu_cfg->clk_src == SOC_CPU_CLK_SRC_XTAL) {
-#endif
-		uart_ll_set_sclk(UART_LL_GET_HW(CONFIG_ESP_CONSOLE_UART_NUM), UART_SCLK_XTAL);
-		uart_ll_set_baudrate(UART_LL_GET_HW(CONFIG_ESP_CONSOLE_UART_NUM),
-				     CONFIG_ESP_CONSOLE_UART_BAUDRATE, cpu_cfg->xtal_freq * MHZ(1));
-		esp_rom_output_tx_wait_idle(CONFIG_ESP_CONSOLE_UART_NUM);
-	}
-#endif
-
-	rtc_clk_cpu_freq_get_config(&old_config);
-
-	ret = rtc_clk_cpu_freq_mhz_to_config(cpu_cfg->cpu_freq, &new_config);
-#if defined(CONFIG_SOC_SERIES_ESP32C5)
-	/*
-	 * C5 DT binding values (ESP32_CPU_CLK_SRC_*) don't match HAL
-	 * enums (SOC_CPU_CLK_SRC_*). Map each DT value to its HAL
-	 * equivalent(s) for validation.
-	 */
-	if (!ret || (cpu_cfg->clk_src == ESP32_CPU_CLK_SRC_PLL
-			     ? (new_config.source != SOC_CPU_CLK_SRC_PLL_F160M &&
-				new_config.source != SOC_CPU_CLK_SRC_PLL_F240M)
-			     : (cpu_cfg->clk_src == ESP32_CPU_CLK_SRC_XTAL
-					? (new_config.source != SOC_CPU_CLK_SRC_XTAL)
-					: (new_config.source != cpu_cfg->clk_src)))) {
-#elif defined(CONFIG_SOC_SERIES_ESP32P4)
-	/*
-	 * P4 uses CPLL as its PLL source.
-	 * The DT binding uses ESP32_CPU_CLK_SRC_PLL (1) as a generic PLL marker.
-	 */
-	if (!ret ||
-	    (cpu_cfg->clk_src == ESP32_CPU_CLK_SRC_PLL ? (new_config.source != SOC_CPU_CLK_SRC_CPLL)
-						       : (new_config.source != cpu_cfg->clk_src))) {
-#else
-	if (!ret || (new_config.source != cpu_cfg->clk_src)) {
-#endif
-		LOG_ERR("invalid CPU frequency value");
-		return -EINVAL;
-	}
-
-#if defined(CONFIG_SOC_SERIES_ESP32C5) || defined(CONFIG_SOC_SERIES_ESP32C6) ||                    \
-	defined(CONFIG_SOC_SERIES_ESP32H2)
-#if defined(CONFIG_SOC_SERIES_ESP32C5)
-	bool keep_pll = (cpu_cfg->clk_src == ESP32_CPU_CLK_SRC_XTAL);
-#else
-	bool keep_pll = (cpu_cfg->clk_src == SOC_CPU_CLK_SRC_XTAL);
-#endif
-
-	if (keep_pll) {
-		rtc_clk_bbpll_add_consumer();
-	}
-#endif
-
-#if defined(CONFIG_SOC_SERIES_ESP32C6)
-	if (cpu_cfg->clk_src == SOC_CPU_CLK_SRC_PLL) {
-		clk_ll_mspi_fast_set_hs_divider(6);
-		/*
-		 * The I2C analog master may be clocked from 160MHz PLL.
-		 * If PLL was disabled, that clock is dead and REGI2C
-		 * writes will hang. Switch to XTAL-based clock before
-		 * re-enabling PLL, then restore 160MHz after PLL is up.
-		 */
-		MODEM_LPCON.i2c_mst_clk_conf.clk_i2c_mst_sel_160m = 0;
-	}
-#endif
-
-	unsigned int key = irq_lock();
-	rtc_clk_cpu_freq_set_config(&new_config);
-
-	esp_cpu_set_cycle_count((uint64_t)esp_cpu_get_cycle_count() * cpu_cfg->cpu_freq /
-				old_config.freq_mhz);
-	irq_unlock(key);
-
-#if defined(CONFIG_SOC_SERIES_ESP32C6)
-	if (cpu_cfg->clk_src == SOC_CPU_CLK_SRC_PLL) {
-		regi2c_ctrl_ll_master_configure_clock();
-	}
-#endif
-
-#if defined(CONFIG_SOC_SERIES_ESP32C5) || defined(CONFIG_SOC_SERIES_ESP32C6) ||                    \
-	defined(CONFIG_SOC_SERIES_ESP32H2)
-	if (keep_pll) {
-		rtc_clk_bbpll_remove_consumer();
-	}
-#endif
-
-#if defined(CONFIG_ESP_CONSOLE_UART)
-#if defined(CONFIG_SOC_SERIES_ESP32C2) || defined(CONFIG_SOC_SERIES_ESP32C5) ||                    \
-	defined(CONFIG_SOC_SERIES_ESP32C6) || defined(CONFIG_SOC_SERIES_ESP32H2) ||                \
-	defined(CONFIG_SOC_SERIES_ESP32P4)
-#if defined(CONFIG_SOC_SERIES_ESP32P4)
-	if (cpu_cfg->clk_src == ESP32_CPU_CLK_SRC_PLL) {
-		/*
-		 * The ROM selects XTAL as the console UART source, and XTAL
-		 * is independent of the CPU/PLL frequency. Keep the console
-		 * on XTAL across the CPU clock change instead of re-sourcing
-		 * it to PLL_F80M: that reference clock is later gated by
-		 * esp_perip_clk_init(), which would leave the console
-		 * without a clock and stall the first transmit.
-		 */
-		uint32_t uart_sclk_freq = (uint32_t)rtc_clk_xtal_freq_get() * MHZ(1);
-
-		uart_ll_set_sclk(UART_LL_GET_HW(CONFIG_ESP_CONSOLE_UART_NUM), UART_SCLK_XTAL);
-		uart_ll_set_baudrate(UART_LL_GET_HW(CONFIG_ESP_CONSOLE_UART_NUM),
-				     CONFIG_ESP_CONSOLE_UART_BAUDRATE, uart_sclk_freq);
-	}
-#else
-#if defined(CONFIG_SOC_SERIES_ESP32C5)
-	if (cpu_cfg->clk_src == ESP32_CPU_CLK_SRC_PLL) {
-#else
-	if (cpu_cfg->clk_src == SOC_CPU_CLK_SRC_PLL) {
-#endif
-		uint32_t uart_sclk_freq;
-
-		esp_clk_tree_src_get_freq_hz(
-			UART_SCLK_DEFAULT, ESP_CLK_TREE_SRC_FREQ_PRECISION_CACHED, &uart_sclk_freq);
-#if defined(CONFIG_SOC_SERIES_ESP32C5)
-		esp_clk_tree_enable_src(UART_SCLK_DEFAULT, true);
-#endif
-		uart_ll_set_sclk(UART_LL_GET_HW(CONFIG_ESP_CONSOLE_UART_NUM), UART_SCLK_DEFAULT);
-		uart_ll_set_baudrate(UART_LL_GET_HW(CONFIG_ESP_CONSOLE_UART_NUM),
-				     CONFIG_ESP_CONSOLE_UART_BAUDRATE, uart_sclk_freq);
-	}
-#endif
-#else
-#if defined(CONFIG_MCUBOOT) && defined(ESP_ROM_UART_CLK_IS_XTAL)
-	uint32_t uart_clock_src_hz = (uint32_t)rtc_clk_xtal_freq_get() * MHZ(1);
-#else
-	uint32_t uart_clock_src_hz = esp_clk_apb_freq();
-#endif
-	uart_ll_set_baudrate(UART_LL_GET_HW(CONFIG_ESP_CONSOLE_UART_NUM),
-			     CONFIG_ESP_CONSOLE_UART_BAUDRATE, uart_clock_src_hz);
-#endif
-#endif
 	return 0;
 }
